feat: replace random connection ID with UUID v7 and widen type to u128

+++ src/models.rs
@@ -1114,7 +1114,7 @@ pub struct ErrorResponse {
     pub message: String,
 }
 
-pub type WsConnectionId = u32;
+pub type WsConnectionId = u128;
 
 #[derive(Default)]
 pub struct ConnectionManager(pub RwLock<HashMap<WsConnectionId, WsWriteKind>>);
